refactor: move to new person package
